fix: do not use link on windows

+++ torappu/core/task/gamedata.py
@@ -3,6 +3,7 @@ import json
 import base64
 import shutil
 import asyncio
+import platform
 import subprocess
 from typing import ClassVar
 
@@ -270,8 +271,10 @@ class GameData(Task):
 
         await asyncio.gather(*(self.client.resolve_ab(ab[:-3]) for ab in gamedata_abs))
         await asyncio.gather(*(self.unpack(ab) for ab in gamedata_abs))
-        STORAGE_DIR.joinpath("asset", "gamedata", "latest").unlink(True)
-        STORAGE_DIR.joinpath("asset", "gamedata", "latest").symlink_to(
-            f"./{self.client.version.res_version}",
-            True,
-        )
+
+        if platform.system() != "Windows":
+            STORAGE_DIR.joinpath("asset", "gamedata", "latest").unlink(True)
+            STORAGE_DIR.joinpath("asset", "gamedata", "latest").symlink_to(
+                f"./{self.client.version.res_version}",
+                True,
+            )
